Update docs to Phase 9 and prototype Phase 10 Hardware Path
- Updated README.md, USER_MANUAL.md, and docs/ to reflect completion of Phases 7-9 (LRU, Async, Telemetry).
- Implemented Phase 10 Hardware Path prototype:
  - Created include/uapi_tfmbs.h with ioctl definitions.
  - Implemented src/tfmbs_driver_mock.c as a kernel-side mock.
  - Refactored src/libtfmbs_device.c into a user-space shim and src/fabric_emulator.c core.
- Added tests/test_phase10.c to verify the new driver abstraction.
- Updated Makefile to include new source files and test targets.
- Verified all tests pass and interposer still functions correctly.
