+++ docs/08_PROFILING.md
@@ -25,7 +25,25 @@ When scaling workloads:
 *   **Weight Broadcast:** Use the `0x9000` address to load weights once and use them across all tiles. This reduces AXI bus congestion.
 *   **Lock-Step:** Tiles share a frame controller. If one tile is masked out, it still consumes "cycles" but zero "utilization" and zero power.
 
-## 4. Best Practices for Performance
+## 4. Phase 9 Telemetry (Interposer)
+
+When using the `libtfmbs_intercept.so` interposer, high-level telemetry is automatically reported to `stderr`. This provides a non-intrusive way to monitor real-world application performance.
+
+### Interpreting the Log
+A typical telemetry output looks like:
+```text
+[TFMBS-Telemetry] GEMV Completed
+  - Zero-Skips: 172,401 (65.8% reduction)
+  - Pool Usage: 84.2 MB / 128 MB (65.8%)
+  - Evictions:  2
+  - Async Queue: 0 in-flight
+```
+
+*   **Zero-Skips:** Total trits skipped across all rows and columns of the GEMV.
+*   **Pool Usage:** Memory occupancy of the 128MB Fabric emulation pool.
+*   **Evictions:** Number of times an LRU block was freed to make room for new data. High eviction counts suggest the model is too large for the 128MB pool, potentially causing "thrashing."
+
+## 5. Best Practices for Performance
 
 1.  **Alignment:** Always try to process trits in multiples of 15. If your vector size is not a multiple of 15, the remaining lanes will be idle.
 2.  **DMA Loading:** For large datasets, use `load_stream()` instead of multiple `load()` calls to leverage the high-bandwidth AXI-Stream path.
