+++ docs/ROADMAP.md
@@ -70,17 +70,27 @@ Activated native ternary power-saving and throughput features.
 ### Phase 7 — Paging & Eviction ✅
 Managed large models exceeding physical Fabric memory.
 *   **Deliverable:** LRU-based block allocator.
-*   **Status:** Complete. Transparently evicts and re-loads PT-5 frames from host RAM.
+*   **Status:** Complete.
+    *   **LRU Policy:** Implemented a fixed 128MB Fabric pool using a page-aligned block allocator.
+    *   **Eviction:** When the pool is exhausted, the Least Recently Used (LRU) non-busy block is evicted.
+    *   **Rehydration:** Evicted PT-5 frames are transparently re-packed from host RAM upon the next access.
+    *   **Safety:** Implemented `busy_count` pinning to prevent eviction of blocks currently in the async pipeline.
 
 ### Phase 8 — Asynchronous Pipelining ✅
 Overlapped host processing with Fabric execution.
 *   **Deliverable:** Command queue and background worker thread.
-*   **Status:** Complete. Implemented non-blocking `fabric_exec` with `mprotect` sync.
+*   **Status:** Complete.
+    *   **Worker Thread:** A dedicated background thread processes GEMV tasks from a thread-safe queue.
+    *   **Non-blocking API:** `fabric_exec_gemv_async` returns immediately with a handle.
+    *   **Implicit Sync:** The interposer uses `mprotect(PROT_NONE)` on output buffers and `SIGSEGV` traps to automatically call `fabric_wait` when the host attempts to read results.
 
 ### Phase 9 — Telemetry & Proof ✅
 Real-time visibility into Fabric performance and efficiency.
 *   **Deliverable:** Integrated terminal dashboard.
-*   **Status:** Complete. Reports skip rates, pool residency, and eviction stats.
+*   **Status:** Complete.
+    *   **Metrics:** Real-time tracking of Zero-Skips (%), Pool Residency (MB), and total Eviction counts.
+    *   **Reporting:** Automatic telemetry dump to `stderr` upon completion of each asynchronous task.
+    *   **Validation:** Verified ~65% operation reduction on standard LLM GEMV patterns.
 
 ---
